+++ resdk/analysis/_register.py
@@ -6,7 +6,7 @@ from resdk.analysis.chip_seq import macs, rose2
 from resdk.analysis.differential_expressions import cuffdiff
 from resdk.analysis.expressions import cuffnorm, cuffquant
 from resdk.analysis.plots import bamliquidator, bamplot
-from resdk.resources import Collection, Relation, Sample
+from resdk.resources import Collection, Data, Relation, Sample
 
 Collection.run_bamliquidator = bamliquidator
 Collection.run_bamplot = bamplot
@@ -17,6 +17,9 @@ Collection.run_cuffquant = cuffquant
 Collection.run_hisat2 = hisat2
 Collection.run_macs = macs
 Collection.run_rose2 = rose2
+
+Data.run_bowtie2 = bowtie2
+Data.run_hisat2 = hisat2
 
 Relation.run_bamliquidator = bamliquidator
 Relation.run_bamplot = bamplot