+++ resdk/tests/unit/test_utils.py
@@ -211,6 +211,21 @@ class TestUtils(unittest.TestCase):
         collection_2._samples = ['sample_2']
         self.assertEqual(get_samples([collection_1, collection_2]), ['sample_1', 'sample_2'])
 
+        data = Data(id=1, resolwe=MagicMock())
+        data._sample = 'sample_1'
+        self.assertEqual(get_samples(data), ['sample_1'])
+
+        data1 = Data(id=1, resolwe=MagicMock())
+        data1._sample = 'sample1'
+        data2 = Data(id=2, resolwe=MagicMock())
+        data2._sample = 'sample2'
+        self.assertEqual(get_samples([data1, data2]), ['sample1', 'sample2'])
+
+        data = Data(id=1, resolwe=MagicMock(**{'sample.filter.return_value': None}))
+        data._sample = None
+        with self.assertRaises(TypeError):
+            get_samples(data)
+
         sample = Sample(id=1, resolwe=MagicMock())
         self.assertEqual(get_samples(sample), [sample])